refactor(pubSubClient): modernize types, constants, and shim headers
- Replace #ifndef guards with #pragma once in all headers
- Convert MQTT state codes to inline constexpr int8_t, packet type
  and QoS constants to inline constexpr uint8_t
- Narrow _state field and state() return type to int8_t
- Change unsigned long timestamps to uint32_t throughout
- Narrow plength from unsigned int to uint16_t in publish/beginPublish
- Narrow publish_P local variables; remove C-style cast
- Add override to ShimClient overrides; drop unnecessary virtual from
  ShimClient, Buffer, and Stream
- Replace 0 with nullptr/false for pointer/bool args in test specs

+++ test/test_pubsubclient_keepalive/keepalive_spec.cpp
@@ -163,7 +163,7 @@ bool test_keepalive_disconnects_hung() {
   }
   IS_FALSE(rc);
 
-  int8_t state = static_cast<int8_t>(client.state());
+  int8_t state = client.state();
   IS_TRUE(state == MQTT_CONNECTION_TIMEOUT);
 
   IS_FALSE(shimClient.error());
